sample_app_components/mcp_server/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
httpx[socks,http2]>=0.27,<0.29
httpx-sse>=0.4.0,<0.5
requests
sseclient-py
azure-cosmos
azure-identity
python-dotenv
aiohttp
dapr-ext-fastapi
isodate
aiohttp
azure-servicebus==7.12.3
psycopg
psycopg_pool
fastmcp
azure-search-documents
azure-ai-evaluation
